+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # walkeros
 
+## 4.3.2
+
+### Patch Changes
+
+- Updated dependencies [a32ac49]
+  - @walkeros/cli@4.3.2
+
 ## 4.3.1
 
 ### Patch Changes