+++ apps/demos/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @walkeros/storybook-demo
 
+## 4.3.2
+
+### Patch Changes
+
+- @walkeros/core@4.3.2
+- @walkeros/web-source-browser@4.3.2
+
 ## 4.3.1
 
 ### Patch Changes