+++ packages/server/stores/sheets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/server-store-sheets
 
+## 4.3.2
+
+### Patch Changes
+
+- @walkeros/core@4.3.2
+
 ## 4.3.1
 
 ### Patch Changes