+++ packages/web/destinations/pinterest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @walkeros/web-destination-pinterest
 
+## 4.3.2
+
+### Patch Changes
+
+- @walkeros/core@4.3.2
+- @walkeros/web-core@4.3.2
+
 ## 4.3.1
 
 ### Patch Changes